+++ src/ShapeCrawler/ShapeCollection/ShapeSize.cs
@@ -17,13 +17,13 @@ internal sealed class ShapeSize
         this.sdkPShapeTreeElement = sdkPShapeTreeElement;
     }
 
-    internal int Height() => UnitConverter.VerticalEmuToPixel(this.AExtents().Cy!);
+    internal decimal Height() => UnitConverter.VerticalEmuToPixel(this.AExtents().Cy!);
    
-    internal void UpdateHeight(int heightPixels) => this.AExtents().Cy = UnitConverter.VerticalPixelToEmu(heightPixels);
+    internal void UpdateHeight(decimal heightPixels) => this.AExtents().Cy = UnitConverter.VerticalPixelToEmu(heightPixels);
     
-    internal int Width() => UnitConverter.HorizontalEmuToPixel(this.AExtents().Cx!);
+    internal decimal Width() => UnitConverter.HorizontalEmuToPixel(this.AExtents().Cx!);
     
-    internal void UpdateWidth(int widthPixels) => this.AExtents().Cx = UnitConverter.HorizontalPixelToEmu(widthPixels);
+    internal void UpdateWidth(decimal widthPixels) => this.AExtents().Cx = UnitConverter.HorizontalPixelToEmu(widthPixels);
 
     private A.Extents AExtents()
     {
